app/pyproject.toml
[project]
name = "rofl-x402"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "agent0-sdk",
    "fastapi>=0.120.2",
    "x402>=0.2.1",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.38.0",
    "coincurve>=20.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
ollama = [
    "ollama>=0.4.4",
]
gaia = [
    "openai>=1.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.pytest.ini_options]
pythonpath = "."
testpaths = ["tests"]
